Refactor firmware code and improve functionality
- Updated CMakeLists.txt to include e100 application source files and directories.
- Added return statements to functions in uboot_env_get.cpp, DSP_mem.c, Sharespace.c, and other files to ensure proper function completion.
- Corrected sprintf usage in ui_api.cpp to dereference the double pointer correctly.
- Changed function return types from void to int in several files for consistency and to provide return values.
- Added missing includes for <cstdint> in various header files to ensure type definitions are available.
- Enhanced ProfileManager class in bed_mesh.cpp to improve profile management functionality.
- Added new header files (cbd_pointer.h, cbd_queue.h) for better modularity and organization.
- Updated toolchain scripts to use wrapper scripts for better compatibility.
- Added .gitignore to exclude build directories from version control.

+++ firmware/core/allwinner/SharespaceInit.h
@@ -90,7 +90,7 @@ int set_arm_ops_addr(char * arm_write_addr,char *dsp_write_addr );
 void sharespace_munmap();
 int sharespace_wait_dsp_init(   );
 int sharespace_read_dsp(uint8_t* buf);
-uint16_t DSP_mem_write(uint8_t* cmd, int len);
+int DSP_mem_write(uint8_t* cmd, int len);
 int DSP_mem_read( uint8_t* buf) ;       //
 void set_DSP_mem_read_pos(uint32_t read_addr);
 void set_arm_mem_sync(  void *fun_sync);
